finish wrap

+++ sponge/tests/wrapping_integers_unwrap.cc
@@ -1,6 +1,7 @@
 #include "test_should_be.hh"
 #include "wrapping_integers.hh"
 
+#include <_types/_uint32_t.h>
 #include <cstdint>
 #include <exception>
 #include <iostream>
@@ -12,19 +13,19 @@ using namespace std;
 int main() {
     try {
         // Unwrap the first byte after ISN
-        test_should_be(unwrap(WrappingInt32(1), WrappingInt32(0), 0), 1ul);
+        test_should_be(unwrap(WrappingInt32(1), WrappingInt32(0), 0), static_cast<uint64_t>(1ul));
         // Unwrap the first byte after the first wrap
-        test_should_be(unwrap(WrappingInt32(1), WrappingInt32(0), UINT32_MAX), (1ul << 32) + 1);
+        test_should_be(unwrap(WrappingInt32(1), WrappingInt32(0), UINT32_MAX), static_cast<uint64_t>((1ul << 32) + 1));
         // Unwrap the last byte before the third wrap
-        test_should_be(unwrap(WrappingInt32(UINT32_MAX - 1), WrappingInt32(0), 3 * (1ul << 32)), 3 * (1ul << 32) - 2);
+        test_should_be(unwrap(WrappingInt32(UINT32_MAX - 1), WrappingInt32(0), 3 * (1ul << 32)), static_cast<uint64_t>(3 * (1ul << 32) - 2));
         // Unwrap the 10th from last byte before the third wrap
-        test_should_be(unwrap(WrappingInt32(UINT32_MAX - 10), WrappingInt32(0), 3 * (1ul << 32)), 3 * (1ul << 32) - 11);
+        test_should_be(unwrap(WrappingInt32(UINT32_MAX - 10), WrappingInt32(0), 3 * (1ul << 32)), static_cast<uint64_t>(3 * (1ul << 32) - 11));
         // Non-zero ISN
-        test_should_be(unwrap(WrappingInt32(UINT32_MAX), WrappingInt32(10), 3 * (1ul << 32)), 3 * (1ul << 32) - 11);
+        test_should_be(unwrap(WrappingInt32(UINT32_MAX), WrappingInt32(10), 3 * (1ul << 32)), static_cast<uint64_t>(3 * (1ul << 32) - 11));
         // Big unwrap
         test_should_be(unwrap(WrappingInt32(UINT32_MAX), WrappingInt32(0), 0), static_cast<uint64_t>(UINT32_MAX));
         // Unwrap a non-zero ISN
-        test_should_be(unwrap(WrappingInt32(16), WrappingInt32(16), 0), 0ul);
+        test_should_be(unwrap(WrappingInt32(16), WrappingInt32(16), 0), static_cast<uint64_t>(0ul));
 
         // Big unwrap with non-zero ISN
         test_should_be(unwrap(WrappingInt32(15), WrappingInt32(16), 0), static_cast<uint64_t>(UINT32_MAX));
